+++ packages/eas-cli/src/graphql/queries/UserQuery.ts
@@ -3,8 +3,13 @@ import gql from 'graphql-tag';
 
 import { ExpoGraphqlClient } from '../../commandUtils/context/contextUtils/createGraphqlClient';
 import { withErrorHandlingAsync } from '../client';
-import { CurrentUserQuery } from '../generated';
+import {
+  AccountFragment,
+  CurrentUserQuery,
+  CurrentUserWithPrimaryAccountQuery,
+} from '../generated';
 import { AccountFragmentNode } from '../types/Account';
+import { MeActorFragmentNode } from '../types/Actor';
 
 export const UserQuery = {
   async currentUserAsync(graphqlClient: ExpoGraphqlClient): Promise<CurrentUserQuery['meActor']> {
@@ -16,31 +21,10 @@ export const UserQuery = {
               meActor {
                 __typename
                 id
-                ... on User {
-                  email
-                }
-                ... on UserActor {
-                  username
-                  primaryAccount {
-                    id
-                    ...AccountFragment
-                  }
-                }
-                ... on Robot {
-                  firstName
-                }
-                accounts {
-                  id
-                  ...AccountFragment
-                }
-                ... on PartnerActor {
-                  username
-                }
-                featureGates
-                isExpoAdmin
+                ...MeActorFragment
               }
             }
-            ${print(AccountFragmentNode)}
+            ${print(MeActorFragmentNode)}
           `,
           {},
           {
@@ -51,5 +35,53 @@ export const UserQuery = {
     );
 
     return data.meActor;
+  },
+  async requireCurrentUserPrimaryAccountAsync(
+    graphqlClient: ExpoGraphqlClient
+  ): Promise<AccountFragment> {
+    let data: CurrentUserWithPrimaryAccountQuery;
+    try {
+      data = await withErrorHandlingAsync(
+        graphqlClient
+          .query<CurrentUserWithPrimaryAccountQuery>(
+            gql`
+              query CurrentUserWithPrimaryAccount {
+                meActor {
+                  __typename
+                  id
+                  ... on UserActor {
+                    primaryAccount {
+                      id
+                      ...AccountFragment
+                    }
+                  }
+                }
+              }
+              
+              ${print(AccountFragmentNode)}
+            `,
+            {},
+            {
+              additionalTypenames: ['User', 'SSOUser'],
+            }
+          )
+          .toPromise()
+      );
+    } catch (error) {
+      throw new Error(
+        'An error occurred while fetching the primary account of the current user. Check to ensure your session has sufficient scope on your primary account.'
+      );
+    }
+
+    const actor = data.meActor;
+    if (!actor) {
+      throw new Error('Must be logged in to perform this action.');
+    }
+
+    if (actor.__typename === 'User' || actor.__typename === 'SSOUser') {
+      return actor.primaryAccount;
+    }
+
+    throw new Error(`This action is not supported for the ${actor.__typename} user type.`);
   },
 };